TicketApp/requirements.txt
asgiref==3.2.10
astroid==2.4.1
cffi==1.14.0
colorama==0.4.3
cryptography==2.9.2
dj-database-url==0.5.0
Django==2.1.15
django-cors-headers==3.4.0
django-heroku==0.3.1
django-pyodbc-azure==2.1.0.0
django-rest-knox==4.1.0
djangorestframework==3.11.0
gunicorn==20.0.4
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2==2.8.5
pycparser==2.20
pylint==2.5.2
pyodbc==4.0.30
pytz==2020.1
six==1.14.0
sqlparse==0.3.1
toml==0.10.1
whitenoise==5.1.0
wrapt==1.12.1